Add older version branch

+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub fn add(left: u64, right: u64) -> u64 {
+pub fn add(left: f64, right: f64) -> f64 {
     left + right
 }
 
@@ -8,7 +8,7 @@ mod tests {
 
     #[test]
     fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
+        let result = add(2.0, 2.0);
+        assert_eq!(result, 4.0);
     }
 }
